auth/requirements.txt
Flask[async]==2.0.2
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
uvicorn==0.17.1
gunicorn==20.1.0
flasgger==0.9.5
psycopg2-binary==2.9.2
pydantic==1.9.0
pytest==7.0.0
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.27.0
sentry-sdk[flask]==1.5.5
backoff==1.11.1
jaeger-client==4.8.0
Flask-OpenTracing==1.1.0
Flask-Caching==1.10.1
redis==4.1.4
authlib==0.15.5
requests==2.27.1
Flask-Migrate==3.1.0
